Add playIfSupported function to useHaptics
